OMG.Output <- OMG.Output.Protocol
