+++ apps/omg/lib/omg/state/transaction/payment.ex
@@ -171,14 +171,14 @@ defimpl OMG.State.Transaction.Protocol, for: OMG.State.Transaction.Payment do
       do: [
         tx_type,
         Enum.map(inputs, &OMG.InputPointer.get_data_for_rlp/1),
-        Enum.map(outputs, &OMG.Output.Protocol.get_data_for_rlp/1),
+        Enum.map(outputs, &OMG.Output.get_data_for_rlp/1),
         # used to be optional and as such was `if`-appended if not null here
         # When it is not optional, and there's the if, dialyzer complains about the if
         0,
         metadata
       ]
 
-  @spec get_outputs(Transaction.Payment.t()) :: list(Output.Protocol.t())
+  @spec get_outputs(Transaction.Payment.t()) :: list(Output.t())
   def get_outputs(%Transaction.Payment{outputs: outputs}), do: outputs
 
   @spec get_inputs(Transaction.Payment.t()) :: list(InputPointer.t())
@@ -202,7 +202,7 @@ defimpl OMG.State.Transaction.Protocol, for: OMG.State.Transaction.Payment do
 
   Returns the fees that this transaction is paying, mapped by currency
   """
-  @spec can_apply?(Transaction.Payment.t(), list(Output.Protocol.t())) ::
+  @spec can_apply?(Transaction.Payment.t(), list(Output.t())) ::
           {:ok, map()} | {:error, :amounts_do_not_add_up}
   def can_apply?(%Transaction.Payment{} = tx, outputs_spent) do
     outputs = Transaction.get_outputs(tx)
