+++ apps/omg/lib/omg/utxo.ex
@@ -23,7 +23,7 @@ defmodule OMG.Utxo do
   defstruct [:output, :creating_txhash]
 
   @type t() :: %__MODULE__{
-          output: Output.Protocol.t(),
+          output: Output.t(),
           creating_txhash: Transaction.tx_hash()
         }
 
@@ -60,7 +60,7 @@ defmodule OMG.Utxo do
   def to_db_value(%__MODULE__{output: output, creating_txhash: creating_txhash})
       when is_nil_or_binary(creating_txhash) do
     %{creating_txhash: creating_txhash}
-    |> Map.put(:output, OMG.Output.Protocol.to_db_value(output))
+    |> Map.put(:output, OMG.Output.to_db_value(output))
   end
 
   def from_db_value(%{output: output, creating_txhash: creating_txhash})
